🏷️(frontend) move payment types into PaymentInterfaces
Types related to Payment were declared into Joanie types. This file is really
too big. With the new sale tunnel a PaymentInterfaces/types module has been
created so it appears a good idea to move all types related to payment in this
module
