+++ src/frontend/js/types/Joanie.ts
@@ -3,7 +3,7 @@ import type { Nullable } from 'types/utils';
 import { Resource, ResourcesQuery } from 'hooks/useResources';
 import { OrderResourcesQuery } from 'hooks/useOrders';
 import { Course as RichieCourse } from 'types/Course';
-import { PaymentErrorMessageId } from 'components/PaymentInterfaces/types';
+import { Payment } from 'components/PaymentInterfaces/types';
 import { JoanieUserProfile } from './User';
 
 // - Generic
@@ -403,44 +403,6 @@ export interface UserWishlistCreationPayload {
 }
 
 // Payment
-export enum PaymentProviders {
-  DUMMY = 'dummy',
-  PAYPLUG = 'payplug',
-  LYRA = 'lyra',
-}
-
-export interface PaymentWithId {
-  payment_id: string;
-}
-
-export interface DummyPayment extends PaymentWithId {
-  provider_name: PaymentProviders.DUMMY;
-  url: string;
-  is_paid?: boolean;
-}
-
-export interface PayplugPayment extends PaymentWithId {
-  provider_name: PaymentProviders.PAYPLUG;
-  url: string;
-  is_paid?: boolean;
-}
-
-export interface LyraPayment {
-  provider_name: PaymentProviders.LYRA;
-  form_token: string;
-  configuration: {
-    public_key: string;
-    base_url: string;
-  };
-}
-
-export type Payment = DummyPayment | PayplugPayment | LyraPayment;
-
-export type PaymentInterfaceProps<P extends Payment = Payment> = P & {
-  onSuccess: () => void;
-  onError: (messageId: PaymentErrorMessageId) => void;
-};
-
 export interface OrderPaymentInfo {
   payment_info: Payment;
 }
